+++ src/components/question/artist/screen.e2e.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import {configure, shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 
-import QuestionArtistScreen from "./question-artist-screen.jsx";
+import QuestionArtistScreen from "./screen.jsx";
 
 configure({adapter: new Adapter()});
